rename schnorr.rs to dleq.rs
